remove new_freshness_policy everywhere
