+++ python_modules/dagster/dagster/_core/definitions/asset_graph.py
@@ -24,7 +24,6 @@ from dagster._core.definitions.declarative_automation.automation_condition impor
 from dagster._core.definitions.events import AssetKey
 from dagster._core.definitions.freshness_policy import FreshnessPolicy
 from dagster._core.definitions.metadata import ArbitraryMetadataMapping
-from dagster._core.definitions.new_freshness_policy import NewFreshnessPolicy
 from dagster._core.definitions.partition import PartitionsDefinition
 from dagster._core.definitions.partition_mapping import PartitionMapping
 from dagster._core.definitions.resolved_asset_deps import ResolvedAssetDependencies
@@ -155,11 +154,6 @@ class AssetNode(BaseAssetNode):
             return {self.key}
         else:
             return self.assets_def.asset_and_check_keys
-
-    @property
-    def new_freshness_policy(self) -> Optional[NewFreshnessPolicy]:
-        """Experimental, do not use."""
-        return self._spec.new_freshness_policy
 
     ##### ASSET GRAPH SPECIFIC INTERFACE
