+++ python_modules/dagster/dagster/_core/definitions/asset_out.py
@@ -17,7 +17,6 @@ from dagster._core.definitions.events import (
 )
 from dagster._core.definitions.freshness_policy import FreshnessPolicy
 from dagster._core.definitions.input import NoValueSentinel
-from dagster._core.definitions.new_freshness_policy import NewFreshnessPolicy
 from dagster._core.definitions.output import Out
 from dagster._core.definitions.partition import PartitionsDefinition
 from dagster._core.definitions.utils import resolve_automation_condition
@@ -210,11 +209,6 @@ class AssetOut:
     def kinds(self) -> Optional[set[str]]:
         return self._spec.kinds
 
-    @property
-    def new_freshness_policy(self) -> Optional[NewFreshnessPolicy]:
-        """Experimental, do not use."""
-        return self._spec.new_freshness_policy
-
     def to_out(self) -> Out:
         return Out(
             dagster_type=self.dagster_type,
@@ -231,7 +225,6 @@ class AssetOut:
         deps: Sequence[AssetDep],
         additional_tags: Mapping[str, str] = {},
         partitions_def: Optional[PartitionsDefinition] = ...,
-        new_freshness_policy: Optional[NewFreshnessPolicy] = ...,
     ) -> AssetSpec:
         return self._spec.replace_attributes(
             key=key,
@@ -239,7 +232,6 @@ class AssetOut:
             kinds=self.kinds,
             deps=[*self._spec.deps, *deps],
             partitions_def=partitions_def if partitions_def is not None else ...,
-            new_freshness_policy=new_freshness_policy if new_freshness_policy is not None else ...,
         )
 
     @public
