+++ python_modules/dagster/dagster/_core/definitions/decorators/asset_decorator.py
@@ -37,7 +37,6 @@ from dagster._core.definitions.events import (
 from dagster._core.definitions.freshness_policy import FreshnessPolicy
 from dagster._core.definitions.input import GraphIn
 from dagster._core.definitions.metadata import ArbitraryMetadataMapping, RawMetadataMapping
-from dagster._core.definitions.new_freshness_policy import NewFreshnessPolicy
 from dagster._core.definitions.output import GraphOut
 from dagster._core.definitions.partition import PartitionsDefinition
 from dagster._core.definitions.policy import RetryPolicy
@@ -91,7 +90,6 @@ def asset(
     owners: Optional[Sequence[str]] = ...,
     kinds: Optional[AbstractSet[str]] = ...,
     pool: Optional[str] = ...,
-    new_freshness_policy: Optional[NewFreshnessPolicy] = ...,
     **kwargs,
 ) -> Callable[[Callable[..., Any]], AssetsDefinition]: ...
 
@@ -120,7 +118,6 @@ def _validate_hidden_non_argument_dep_param(
 @beta_param(param="resource_defs")
 @beta_param(param="io_manager_def")
 @beta_param(param="backfill_policy")
-@beta_param(param="new_freshness_policy")
 @hidden_param(
     param="non_argument_deps",
     breaking_version="2.0.0",
@@ -170,7 +167,6 @@ def asset(
     owners: Optional[Sequence[str]] = None,
     kinds: Optional[AbstractSet[str]] = None,
     pool: Optional[str] = None,
-    new_freshness_policy: Optional[NewFreshnessPolicy] = None,
     **kwargs,
 ) -> Union[AssetsDefinition, Callable[[Callable[..., Any]], AssetsDefinition]]:
     """Create a definition for how to compute an asset.
@@ -253,7 +249,7 @@ def asset(
         pool (Optional[str]): A string that identifies the concurrency pool that governs this asset's execution.
         non_argument_deps (Optional[Union[Set[AssetKey], Set[str]]]): Deprecated, use deps instead.
             Set of asset keys that are upstream dependencies, but do not pass an input to the asset.
-            Hidden parameter not exposed in the decorator signature, but passed in kwargs. Currently experimental.
+            Hidden parameter not exposed in the decorator signature, but passed in kwargs.
 
     Examples:
         .. code-block:: python
@@ -329,7 +325,6 @@ def asset(
         key=key,
         owners=owners,
         pool=pool,
-        new_freshness_policy=new_freshness_policy,
     )
 
     if compute_fn is not None:
@@ -403,7 +398,6 @@ class AssetDecoratorArgs(NamedTuple):
     check_specs: Optional[Sequence[AssetCheckSpec]]
     owners: Optional[Sequence[str]]
     pool: Optional[str]
-    new_freshness_policy: Optional[NewFreshnessPolicy]
 
 
 class ResourceRelatedState(NamedTuple):
@@ -529,7 +523,6 @@ def create_assets_def_from_fn_and_decorator_args(
             decorator_name="@asset",
             execution_type=AssetExecutionType.MATERIALIZATION,
             pool=args.pool,
-            new_freshness_policy=args.new_freshness_policy,
         )
 
         builder = DecoratorAssetsDefinitionBuilder.from_asset_outs_in_asset_centric_decorator(
@@ -709,7 +702,6 @@ def multi_asset(
         decorator_name="@multi_asset",
         execution_type=AssetExecutionType.MATERIALIZATION,
         pool=pool,
-        new_freshness_policy=None,
     )
 
     def inner(fn: Callable[..., Any]) -> AssetsDefinition:
