+++ python_modules/dagster/dagster/_core/definitions/remote_asset_graph.py
@@ -37,7 +37,6 @@ from dagster._core.definitions.declarative_automation.automation_condition impor
 )
 from dagster._core.definitions.freshness_policy import FreshnessPolicy
 from dagster._core.definitions.metadata import ArbitraryMetadataMapping
-from dagster._core.definitions.new_freshness_policy import NewFreshnessPolicy
 from dagster._core.definitions.partition import PartitionsDefinition
 from dagster._core.definitions.partition_mapping import PartitionMapping
 from dagster._core.definitions.utils import DEFAULT_GROUP_NAME
@@ -128,11 +127,6 @@ class RemoteAssetNode(BaseAssetNode, ABC):
             else []
         )
 
-    @property
-    def new_freshness_policy(self) -> Optional[NewFreshnessPolicy]:
-        """Experimental, do not use."""
-        return self.resolve_to_singular_repo_scoped_node().asset_node_snap.new_freshness_policy
-
 
 @whitelist_for_serdes
 @record
@@ -204,11 +198,6 @@ class RemoteRepositoryAssetNode(RemoteAssetNode):
     @property
     def pools(self) -> Optional[set[str]]:
         return self.asset_node_snap.pools
-
-    @property
-    def new_freshness_policy(self) -> Optional[NewFreshnessPolicy]:
-        """Experimental, do not use."""
-        return self.asset_node_snap.new_freshness_policy
 
 
 @whitelist_for_serdes
