+++ python_modules/dagster/dagster_tests/core_tests/host_representation_tests/test_external_data.py
@@ -24,9 +24,9 @@ from dagster._core.definitions.asset_spec import AssetExecutionType, AssetSpec
 from dagster._core.definitions.backfill_policy import BackfillPolicy
 from dagster._core.definitions.definitions_class import Definitions
 from dagster._core.definitions.external_asset import external_assets_from_specs
+from dagster._core.definitions.freshness import TimeWindowFreshnessPolicy
 from dagster._core.definitions.metadata import MetadataValue, TextMetadataValue, normalize_metadata
 from dagster._core.definitions.multi_dimensional_partitions import MultiPartitionsDefinition
-from dagster._core.definitions.new_freshness_policy import TimeWindowFreshnessPolicy
 from dagster._core.definitions.partition import ScheduleType
 from dagster._core.definitions.time_window_partitions import TimeWindowPartitionsDefinition
 from dagster._core.definitions.utils import DEFAULT_GROUP_NAME
@@ -237,8 +237,8 @@ def test_asset_invalid_group_name():
             return 1
 
 
-def test_asset_spec_with_new_freshness_policy():
-    """Can we define an asset with a new freshness policy?"""
+def test_asset_spec_with_internal_freshness_policy():
+    """Can we define an asset with an internal freshness policy?"""
     asset1 = AssetSpec(
         key=AssetKey("asset1"),
         internal_freshness_policy=TimeWindowFreshnessPolicy(time_window_minutes=10),
@@ -250,18 +250,6 @@ def test_asset_spec_with_new_freshness_policy():
     assert isinstance(policy, TimeWindowFreshnessPolicy)
     assert policy.time_window_minutes == 10
     assert policy.warning_time_window_minutes is None
-
-
-def test_asset_with_new_freshness_policy():
-    @asset(new_freshness_policy=TimeWindowFreshnessPolicy(time_window_minutes=10))
-    def asset1():
-        return 1
-
-    asset_node_snaps = _get_asset_node_snaps_from_definitions(Definitions(assets=[asset1]))
-
-    assert asset_node_snaps[0].new_freshness_policy == TimeWindowFreshnessPolicy(
-        time_window_minutes=10
-    )
 
 
 def test_two_asset_job():
